Python 3.4 compatible with PEP8 compliance
Updated the code to work with Python 3.4 (print statement parenthesis
and range (changed py2.7 xrange to range)). Edited the code formatting
to align more with PEP8, albeit I did not change the nomenclature of the
matrix objects (i.e. A, B, and C), as I think that would alter the
structure too much.